Deprecate obsolete temporal classes.
