+++ phenol-annotations/src/main/java/org/monarchinitiative/phenol/annotations/base/TemporalRangeAware.java
@@ -1,5 +1,6 @@
 package org.monarchinitiative.phenol.annotations.base;
 
+@Deprecated
 public interface TemporalRangeAware {
 
   TemporalRange temporalRange();
